refactor: move serializer into an other code file
