+++ synth-converter/src/graph/mod.rs
@@ -1,3 +1,4 @@
 pub mod graph_builder;
 pub mod namespaces;
 pub mod utils;
+pub mod rdf_serializers;
